Time: 3 ms (16.96%), Space: 8 MB (41.04%) - LeetHub

+++ 0278-first-bad-version/0278-first-bad-version.cpp
@@ -4,7 +4,7 @@
 class Solution {
 public:
     int firstBadVersion(int n) {
-        long long low=1,high=n;
+        int low=1,high=n;
         while(low<=high){
             long long mid=low+(high-low)/2;
             if(isBadVersion(mid)){
